New js structure with Vue
