+++ package.json
@@ -1,25 +1,28 @@
 {
   "name": "cachet",
   "devDependencies": {
-    "gulp": "~3.9.1",
-    "laravel-elixir": "~6.0.0-9",
     "animate-sass": "git+https://github.com/tgdev/animate-sass.git",
     "autosize": "^3.0.15",
-    "bootstrap-sass": "^3.3.6",
+    "bootstrap-sass": "^3.3.7",
     "chart.js": "^2.2.0",
     "eonasdan-bootstrap-datetimepicker": "~4.17",
     "github-markdown-css": "^2.3.0",
+    "gulp": "~3.9.1",
     "ionicons": "~2.0",
+    "jquery": "~2.2",
     "jquery-minicolors": "^2.1.10",
     "jquery-serializeobject": "^1.0.0",
     "jquery-sparkline": "^2.3.2",
-    "jquery": "~2.2",
+    "laravel-elixir": "~6.0.0-11",
+    "laravel-elixir-webpack-official": "^1.0.6",
     "livestamp": "git+https://github.com/mattbradley/livestampjs.git#develop",
     "lodash": "^4.12.0",
     "messenger": "git+https://github.com/HubSpot/messenger.git",
     "moment": "^2.13.0",
     "sortablejs": "^1.4.2",
-    "sweetalert2": "^4.0.4"
+    "sweetalert2": "^4.0.4",
+    "vue": "^2.0.3",
+    "vue-resource": "^1.0.3"
   },
   "private": true
 }
